Fallback Shift Working

+++ packages/0confirmation-swap-react/src/lib/persistence.js
@@ -5,13 +5,15 @@ import LiquidityRequestParcel from '@0confirmation/sdk/liquidity-request-parcel'
 
 export const saveLoan = (loan) => {
   try {
-    const i = String(Number(localStorage.getItem("index") || -1) + 1);
+    const isAlreadySaved = loan.localIndex;
+    const i = loan.localIndex || String(Number(localStorage.getItem("index") || -1) + 1);
     loan.localIndex = i;
     localStorage.setItem(
       i,
       JSON.stringify({
-        state: loan.utxo ? 'deposited' : 'signed',
+        state: loan.state === 'forced' ? loan.state : loan.utxo ? 'deposited' : 'signed',
         token: loan.token,
+        isReady: loan.isReady,
         amount: String(loan.amount),
         gasRequested: String(loan.gasRequested),
         shifterPool: loan.shifterPool,
@@ -22,7 +24,7 @@ export const saveLoan = (loan) => {
         utxo: loan.utxo
       })
     );
-    localStorage.setItem("index", i);
+    if (!isAlreadySaved) localStorage.setItem("index", i);
   } catch (e) {
     // eslint-disable-line
   }
@@ -40,16 +42,20 @@ export const loadLoan = (i, zero) => {
         gasRequested: parsed.gasRequested,
         signature: parsed.signature,
         nonce: parsed.nonce,
+        amount: parsed.amount,
         actions: parsed.actions,
         forbidLoan: parsed.forbidLoan
       });
+      parcel.isReady = parsed.isReady;
+      parcel.state = 'signed';
       parcel.localIndex = String(i);
       return parcel;
-    } else if (parsed.state === 'deposited') {
+    } else if (parsed.state === 'deposited' || parsed.state === 'forced') {
       const parcel = new DepositedLiquidityRequestParcel({
         zero,
         shifterPool: parsed.shifterPool,
         token: parsed.token,
+        amount: parsed.amount,
         gasRequested: parsed.gasRequested,
         signature: parsed.signature,
         nonce: parsed.nonce,
@@ -57,7 +63,10 @@ export const loadLoan = (i, zero) => {
         forbidLoan: parsed.forbidLoan,
         utxo: parsed.utxo
       });
+      parcel.isReady = parsed.isReady;
       parcel.localIndex = i;
+      parcel.state = parsed.state;
+      return parcel;
     }
   } catch (e) {
     return null;
@@ -66,9 +75,16 @@ export const loadLoan = (i, zero) => {
 
 export const loadLoans = (zero) => {
   const index = Number(localStorage.getItem('index'));
-  return Array(index + 1).fill(0).map((v, i) => loadLoan(zero, i));
+  return Array(index + 1).fill(0).map((v, i) => loadLoan(i, zero));
 };
 
 export const removeLoan = (i) => {
   localStorage.removeItem(String(i));
+};
+
+window.clearLoansDebug = (v) => {
+  for (let i = 0; i < Number(localStorage.getItem('index')); i++) {
+    localStorage.removeItem(String(i));
+  }
+  localStorage.removeItem('index');
 };
